services/api-gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
redis==5.0.7
sqlalchemy==2.0.32
psycopg[binary]==3.2.10
sse-starlette==2.1.0
python-dotenv==1.0.1
pyjwt==2.9.0
bcrypt==4.2.0
prometheus-client==0.20.0
